Be more POSIXy w.r.t. mktemp usage.
If /bin/sh is busybox, and it provides a mktemp builtin, you cannot even
get the default GNU mktemp which is available on all Linux systems like
ever. And surprise, busybox mktemp does not support `--suffix`.

Using a whole temporary directory just to add the .md suffix is sorta
overkill, but OTOH the filename prefix is also now legible, so I guess
this balances out.